C#: Factor C++ parts out of autobuilder
