+++ csharp/autobuilder/Semmle.Autobuild.CSharp/XmlBuildRule.cs
@@ -1,0 +1,21 @@
+﻿using Semmle.Autobuild.Shared;
+
+namespace Semmle.Autobuild.CSharp
+{
+    /// <summary>
+    /// XML extraction.
+    /// </summary>
+    class XmlBuildRule : IBuildRule
+    {
+        public BuildScript Analyse(Autobuilder builder, bool auto)
+        {
+            if (!builder.Options.Indexing || builder.Odasa is null)
+                return BuildScript.Success;
+
+            var command = new CommandBuilder(builder.Actions).
+                RunCommand(builder.Odasa).
+                Argument("index --xml --extensions config csproj props xml");
+            return command.Script;
+        }
+    }
+}
